Move weekProperties et co. to separate module
